MKT_Labs_parking_and_web_vFinal/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
eventlet==0.36.1
pyserial==3.5
